pyproject.toml
[tool.poetry]
name = "p1-ddb"
version = "0.0.1"
description = "Uni stuff"
authors = ["ImMamey <32584037+ImMamey@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "p1_ddb"}]

[tool.poetry.dependencies]
python = "^3.10"
python-dateutil = "^2.8.2"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
mypy = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
